+++ lib/src/models/package_settings.dart
@@ -21,6 +21,13 @@ class PackageSettings {
   /// A list of languages to generate.
   final List<SupportedLanguage>? languagesToGenerate;
 
+  /// A seed for random generators.
+  ///
+  /// This should ensure that psuedo strings are generated identically each run.
+  ///
+  /// Note that new releases of dart may update random seed algorithm.
+  final int seed;
+
   /// Whether the pseudo text should be wrapped in square brackets. Defaults to `true`.
   final bool useBrackets;
 
@@ -48,6 +55,7 @@ class PackageSettings {
     required bool? replaceBase,
     required List<String>? unicodeBlocks,
     required List<String>? languagesToGenerate,
+    required int? seed,
     required bool? useBrackets,
     required String? textExpansionFormat,
     required this.textExpansionRatio,
@@ -62,6 +70,7 @@ class PackageSettings {
             : null,
         languagesToGenerate =
             Utils.covertSupportedLanguagesFromListString(languagesToGenerate),
+        seed = seed ?? DefaultSettings.seed,
         useBrackets = useBrackets ?? DefaultSettings.useBrackets,
         textExpansionFormat =
             Utils.convertTextExpansionFormatFromString(textExpansionFormat) ??
@@ -70,8 +79,8 @@ class PackageSettings {
         arbSettings = arbSettings ?? ARBSettings(),
         csvSettings = csvSettings ?? CSVSettings();
 
-  /// Returns a String representation of the model.
   @override
-  String toString() =>
-      '{inputFilepath: $inputFilepath, replaceBase: $replaceBase, unicodeBlocks, $unicodeBlocks, languagesToGenerate: $languagesToGenerate, useBrackets: $useBrackets, textExpansionFormat: $textExpansionFormat, textExpansionRatio: $textExpansionRatio, patternToIgnore: $patternToIgnore, keysToIgnore: $keysToIgnore, arbSettings: $arbSettings, csvSettings: $csvSettings}';
+  String toString() {
+    return 'PackageSettings(inputFilepath: $inputFilepath, replaceBase: $replaceBase, unicodeBlocks: $unicodeBlocks, languagesToGenerate: $languagesToGenerate, seed: $seed, useBrackets: $useBrackets, textExpansionFormat: $textExpansionFormat, textExpansionRatio: $textExpansionRatio, patternToIgnore: $patternToIgnore, keysToIgnore: $keysToIgnore, arbSettings: $arbSettings, csvSettings: $csvSettings)';
+  }
 }